requirements.txt
albumentations==1.3.0
einops==0.6.0
fastapi==0.89.1
invisible-watermark>=0.1.5
numpy==1.23.1
cudatoolkit==11.3
omegaconf==2.3.0
Pillow==9.4.0
pydantic==1.10.4
pytorch_lightning==1.9.0
requests==2.28.1
scipy==1.9.1
timm==0.6.12
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.0
xformers==0.0.16
opencv-python==4.6.0.66
